+++ src/main/java/com/portfolio/BackEnd/Interface/IHabilidadService.java
@@ -9,7 +9,7 @@ public interface IHabilidadService {
 
     public void saveHabilidad(Habilidad habilidad);
 
-    public void deleteHabilidad(int id);
+    public void deleteHabilidad(Long id);
 
-    public Habilidad findHabilidad(int id);
+    public Habilidad findHabilidad(Long id);
 }
