+++ src/main/java/com/portfolio/BackEnd/Interface/IPersonaService.java
@@ -5,11 +5,16 @@ import java.util.List;
 
 public interface IPersonaService {
 
+     //Traer una lista de personas
     public List<Persona> getPersona();
-
+    
+    //guardar un objeto tipo persona
     public void savePersona(Persona persona);
-
-    public void deletePersona(int id);
-
-    public Persona findPersona(int id);
+    
+    //Eliminar un objeto pero lo buscamos por ID
+    public void deletePersona(Long id);
+    
+    
+    //Buscar una persona por id
+    public Persona findPersona(Long id);
 }
