+++ src/main/java/com/portfolio/BackEnd/Repository/IContactoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface IContactoRepository extends JpaRepository <Contacto,Integer>{
+public interface IContactoRepository extends JpaRepository <Contacto,Long>{
     
 }
