+++ src/main/java/com/portfolio/BackEnd/Repository/IPersonaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPersonaRepository extends JpaRepository <Persona,Integer>{
+public interface IPersonaRepository extends JpaRepository <Persona,Long>{
     
 }
