+++ src/main/java/com/portfolio/BackEnd/Repository/IProyectoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IProyectoRepository extends JpaRepository <Proyecto,Integer>{
+public interface IProyectoRepository extends JpaRepository <Proyecto,Long>{
     
 }
