+++ src/main/java/com/portfolio/BackEnd/Repository/IResumenTrabRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IResumenTrabRepository extends JpaRepository <ResumenTrab,Integer> {
+public interface IResumenTrabRepository extends JpaRepository <ResumenTrab,Long> {
     
 }
