+++ src/main/java/com/portfolio/BackEnd/Service/ImpPersonaService.java
@@ -25,12 +25,12 @@ public class ImpPersonaService implements IPersonaService {
     }
 
     @Override
-    public void deletePersona(int id) {
+    public void deletePersona(Long id) {
         ipersonaRepository.deleteById(id);
     }
 
     @Override
-    public Persona findPersona(int id) {
+    public Persona findPersona(Long id) {
         Persona persona = ipersonaRepository.findById(id).orElse(null);
         return persona;
     }
